requirements.txt
FastAPI==0.109.2
uvicorn==0.27.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
